NVIDIA: Enable relocatable SEC module
Fork the ArmPkg PrePi module to allow for relocation of the module as
the system may load UEFI firmware at any DRAM address.

The CPUBL carveout is used for the initial UEFI memory region with the
stack for the primary CPU at the end of this memory region.

The main FV is detected by looking for the _FVH signature at 64KiB
boundaries.

FD structures in fdf have been removed and FVMAIN_COMPACT firmware image
has been renamed to UEFI_NS. The size of the firmware volume is now the
size of the actual modules in it.

Jira TEGRAUEFI-398

Change-Id: I009137705f8291c34227fd7bbe3cc2d76d5106ce
Reviewed-on: https://git-master.nvidia.com/r/1964648
Tested-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>