Silicon/NVIDIA: Move FV based DT lookup to PrePi
The PrePiLib needed for reading from the FV does not support Dxe, so
SystemResourceLib is could not be used for that. Move that function to
general PrePi code. This code would do nothing for cases where there is
no device tree file in the main uncompressed firmware volume.

Change-Id: I890825f936819d8f1d95ae4357e5b02b765d2e68
Reviewed-on: https://git-master.nvidia.com/r/2026110
Tested-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>