Silicon/NVIDIA: Move Hob list to largest region
Move the Hob list to hardest region when memory is detected to allow for
larger space. This is needed for code coverage enabled builds as the
build size is larger and won't fit in the main region.

This re-adds this change that was reverted due to USB issues, these
issues no longer appear to occur and there was a change to better handle
memory usage that was allocated out of original HOB list

Change-Id: I8115df175edb4a404a4330311291f165bff9b7e8
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2464113
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>