Silicon/NVIDIA: Allow for seperate console and debug output
For T194 when direct UART access is enabled (Sidecar builds) move the
console to a different serial port (uartb). Add dependency between
serial port clock enablement and when the SerialDxe driver can load so
the reduction of the clock can be done prior to the initialization of
the console instead of at exit boot services.

Change-Id: I64b01dd4cb47917ffc54c73599ece787bf0da7c1
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2317926
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>