Platform/NVIDIA: Enable Secure Boot
This change addes Secure Boot menu to UEFI GUI which which
we can turn on and off secure boot. It also adds library to
authenticate and verify an image once UEFI secure boot is turned
on. UEFI secure boot is not turned on by default

Jira TEGRAUEFI-358

Change-Id: If5c788a80a6bd205405e45ccfcf260553e40b59b
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2342299
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit