Silicon/NVIDIA/Drivers/EqosDeviceDxe: Enable for Orin
Enable GPIO and SeRng to run on Orin SIM and use SeRng to get a random
MAC address instead of reading from EEPROM which is not existent in SIM.

Orin DT is not mature and the node status of GPIO and SeRNG is disabled.
They are overridden and populated for the presil platform.

TEGRAUEFI-225

Change-Id: Icf128619ae4e2aafa5068ada8f1f32c256f20ef1
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2353040
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Irene Park <ipark@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>