NVIDIA: FVB Read/Write Interface Fixes
Fix the FvbRead and FvbWrite functions to return EFI_BAD_BUFFER_SIZE
instead of EFI_SUCCESS if a read or write is shortened because it
would cross a block boundary. This is the behavior specified
by the UEFI spec.

Also updates other cases that return EFI_BAD_BUFFER_SIZE to set
the number of bytes read or written to 0 for consistency.

Jira TEGRAUEFI-563

Change-Id: I3e6ad2744e37ae44c4c34596af698695a3be60b0
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2426808
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Porter Jones <porterj@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Porter Jones <porterj@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>