NVIDIA: Update Readme to indicate a specific Linux version
Update the instructions for booting a mainline Linux kernel to use
a specific Linux kernel version so that it is clear what we have
tested with and what is known to work. Other versions may also work
but it is best to indicate a known working version to users to try.

Bug 2877970

Change-Id: Ic99c87954e410f45a3b9755f7f8c2e494721b6f5
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2441916
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Jonathan Hunter <jonathanh@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>