NVIDIA: Add Target Specific MMU MMIO Support
Add support for being able to add target specific MMIO regions
in the MMU memory map. This becomes useful now that we have to
use some HW registers that are not present on all targets and
are not present in DT either for example MCE ARI space.

Change-Id: I05f22cd6cfb92afeeb07b3f9fe11e558fa0c9e42
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2509708
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Swatisri Kantamsetti <swatisrik@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit