NVIDIA/Jetson: Update the details for PCIe support with ACPI
The changes required for supporting PCIe on Tegra194 with ACPI have
been merged for the Linux v5.13 release. Update the details in the UEFI
README to indicate that it is only necessary to patch kernels prior to
v5.13 and provide links to the final patches that were merged.

Finally, update the README to make it clear that it is only necessary
to explicitly enable PCIe support when using ACPI.

Bug 200730107

Change-Id: I3fa66b973e96516679ac7269e0ef079b9c2c7d0d
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2550811
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Eric Brower <ebrower@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Jonathan Hunter <jonathanh@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>