NVIDIA: Disable Interrupts for Orin
Force GIC driver on GicV3 to use Gicv2 legacy mode which is actually
absent in the core so it, in turn, does not enable GIC on Orin. Because
there are no interrupts, the BDS timer before launching the boot option
does not work so the timeout for that has been changed to 0s.

Change has no impact on Xavier.

Change-Id: Ied61b014de8230e0747c10353c373c35435b7a7e
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2-platforms/+/2606065
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>