feat: ADD StMM Nor flash drivers
Add the NOR-Flash drivers for the StMM (Standalone MM) image.
This driver is used by the FVB driver in the StMM image to access
the SPI-NOR media.

The big difference between the Non-StMM driver and this driver is
that the StMM doesn't support the driver binding protocol since there
isn't DT support on StMM. Also on platforms that run OP-TEE the MMIO
base address for the QSPI controller is passed up by OP-TEE to StMM.

Signed-off-by: Mayur Gudmeti <mgudmeti@nvidia.com>
Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>