Silicon/NVIDIA/Drivers: ADD StMM Nor flash drivers
Add the NOR-Flash drivers for the StMM (Standalone MM) image.
This driver is used by the FVB driver in the StMM image to access
the SPI-NOR media.

The big difference between the Non-StMM driver and this driver is
that the StMM doesn't support the driver binding protocol since there
isn't DT support on StMM. Also on platforms that run OP-TEE the MMIO
base address for the QSPI controller is passed up by OP-TEE to StMM.

JIRA TEGRAUEFI-703
Signed-off-by: Mayur Gudmeti <mgudmeti@nvidia.com>
Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Change-Id: I94bcf8ca497eb7dd9a3f9b2e15a9ed78cad6e1eb
Reviewed-on: https://git-master.nvidia.com/r/c/tegra/bootloader/uefi/edk2-nvidia/+/2677733
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: svc-sw-mobile-l4t <svc-sw-mobile-l4t@nvidia.com>
GVS: Gerrit_Virtual_Submit