feat(gop): GOP driver build enhancements
Configure Jetson to build using a binary gop driver, but introduce
overrides so we can build from source.  The intent is to keep any
specific references to internal source directories or build processes
out of edk2-nvidia.

While edk2-nvidia will be open sourced, we will not be distributing the
source for the gop driver.  Instead, we'll distribute a pre-built,
binary efi of the gop driver.  This hybrid source/binary model is common
in edk2.  The binaries are usually distributed in "non-osi" repos.  As
such, we will be introducing "edk2-nvidia-non-osi".

Internally, we'll always build the gop driver from source, but
externally, we'll always use the binary.  Since Jetson's
PlatformBuild.py will be distributed externally, it must be set up to
use "edk2-nvidia-non-osi" and to reference the binary.  When we build
internally, we'll pass in overrides to trigger a source build.

Specifics:
- Use a NvGopDriver.dsc.inc NvGopDriver.fdf.inc to include the driver.
  There will be an internal and external version of these files.
- Back out the temporary GetPackagesPath() and GetRequiredSubmodules()
  added to Jetson/PlatformBuild.py.  These reference internal repos to
  build gop from source.  We'll use the new command-line options
  instead.
- Add --insert-packages-path and --require-submodule command-line
  options as alternatives to GetPackagesPath() and
  GetRequiredSubmodules().  Our internal builds will leverage these
  options.
- Generate images/builddir_<plat>_<tgt>.txt.  The contents of this file
  is the variant's build directory.  This avoids having to make
  assumptions about the build directory's naming convention when digging
  for efi files or other build artifacts.

Bug 3563003

Change-Id: Ib292d2145ab898d5b9458c5b11f93da460ec3ff9
Reviewed-on: https://git-master.nvidia.com/r/c/tegra/bootloader/uefi/edk2-nvidia/+/2683819
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc-sw-mobile-l4t <svc-sw-mobile-l4t@nvidia.com>
Reviewed-by: Jan Bobek <jbobek@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>
GVS: Gerrit_Virtual_Submit
(cherry picked from commit 54ce3f32847f3a78123d2f518a1669a091061788)
Reviewed-on: https://git-master.nvidia.com/r/c/tegra/bootloader/uefi/edk2-nvidia/+/2686049
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>