chore: add dtb overlay for rootfs information
Add dtb overlays to initialize the rootfs information:

L4TRootfsInfo.dts: general rootfs info without A/B enabled.
It is used for general case.

L4TRootfsABInfo.dts: rootfs A/B is enabled and booting with
rootfs A. It is used for rootfs A/B enabled case.

L4TRootfsBrokenInfo.dts: rootfs is broken and booting with
recovery image. It is used for image-based OTA.

Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Wei Ni <wni@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>