feat: Variable based default boot order support
Add support for setting default boot order via a variable.
Currently supports class type options.
Future enhancement will add support for specifing the instance
of a device.

Variable that controls this is DefaultBootPriority and contains
a comma seperated string with the following values.

  "scsi"     - SCSI devices
  "usb"      - USB devices
  "sata"     - SATA devices
  "pxev4"    - IPv4 PXE boot
  "httpv4"   - IPv4 HTTP boot
  "pxev6"    - IPv6 PXE boot
  "httpv6"   - IPv6 HTTP boot
  "nvme"     - Nvme devices
  "ufs"      - UFS devices
  "sd"       - SD devices
  "emmc"     - eMMC devices
  "cdrom"    - CD/DVD drives
  "boot.img" - Android style boot.img partition

Any entries that are not in the list will be left at the end of the
boot order.

If the variable is not found "boot.img,usb,sd,emmc,ufs" is used

Bug 3725156

Change-Id: Iba37e5985d1dfb1139bac7f10f221a2376a5bc99
Signed-off-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/tegra/bootloader/uefi/edk2-nvidia/+/2747710
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc-sw-mobile-l4t <svc-sw-mobile-l4t@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>