fix(pcie): Reduce AXI slave timeout value
Decrease the AXI slave timeout value to 7ms to be in line with
the CBB logic's timeout value and to avoid CBB reporting errors
because of no response from the PCIe IPs AXI slave logic for
configuration space accesses through ECAM when the PCIe link is down.
Also, set the Completion Timeout value to Range-A: 1ms~10ms to be inline
with the AXI  timeout value.

Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Vidya Sagar <vidyas@nvidia.com>