feat: Add DICE certificates reading support
This patch adds codes to read the active DICE ECA certificate from the
WORM partition.

The WORM partition is able to store at most "MM_DICE_CERT_NUM_MAX"
certificates. The data format of each block is:

<Magic, 8 bytes><Type, 4 bytes><Length, 4 bytes><Certificate data>

We will read from the last block and return the first valid certificate
data. In this way, the clients don't need to specify which certificate
it wants.

Signed-off-by: Mark Zhang <markz@nvidia.com>
Reviewed-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>