fix: correct eeprom customer part number handling
The logic in board eeprom handling does not work correctly for
customer part numbers, as it did with previous implementations.

This change fixes the addition of board ids with the leading 0xcc
byte to the /chosen/ids node. It also fixes comparison logic for
applying overlays based on these board ids.

Signed-off-by: Kurt Kiefer <kekiefer@gmail.com>