fix: gIpmiTransportProtocolGuid is not installed
After TegraI2cDxe switched to DirectEnumeration support recently,
SSIF has been broken This is because the SSIF driver callback is
running on a protocol install that occurs before the SSIF device
has been added to the I2C map.

Prior to DirectEnumeration, the callback occurred later, masking
the issue.

The fix is to install gEfiI2cMasterProtocolGuid,
gEfiI2cEnumerateProtocolGuid, and
gEfiI2cBusConfigurationManagementProtocolGuid
once the I2c init is completed.

Signed-off-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>