fix(build): Require at least stuart 0.17.1
Force at least the 0.17.1 version of edk2-pytool-extensions, which
contains the stuart build system.

Currently, pip-requirements.txt requires just 0.16.x of
edk2-pytool-extensions.  However that version contains a bug that
results in random reordering of PACKAGES_PATH, making it impossible to
use "--insert-packages-path" as intended.  The bug was fixed in 0.17.1.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>