feat: add qspi driver to standalonemm
Add a new QSPI driver for StandaloneMM. The driver is a minimal
driver that doesn't do resource voting (assumption is that the
earlier BL setup the controller resources). Make mods to the
NorFlash driver to look for the QSPI controller protocol rather
than use the QspiLib directly.

Add this new driver to the Optee Standalone Mm image.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>