fix(build): Fix stuart to version 0.17.1
Make sure we use a known-good version of edk2-pytool-extensions, which
is currently 0.17.1, instead of allowing pip to upgrade to the latest.

The 0.20.0 version of edk2-pytool-extensions pulls in the 0.12.x of
edk2-pytool-library, which has code that requires Python 3.9 (although
the dependency is not identified in the package metadata).  We want to
maintain compatibility with older versions of Ubuntu which will have
older versions of Python3.

Signed-off-by: Jake Garver <jake@nvidia.com>