fix(stmm): allow support for slow reads
fast reads can't be supported if the controller frequency is below
a certain threshold (usually 50Mhz). In such scenarios we could end
up with errors when trying to read-back data.

Ideally we want to be able to dynamically detect the controller speed
and make this switch, but use a PCD for now and set this on a per
platform as we don't have the clock libraries in StandaloneMm yet.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>