feat: Enable report generation for HostBasedTests
Set BUILDREPORTING=TRUE before going into the build.  This triggers
UefiBuilder to add the appropriate arguments to build.py and generate
the report.

This had to be shoehorned into our settings init() because ci_build
doesn't offer an opportunity to use a UefiBuilder subclass. Fortunately,
we can still do it in our NVIDIACiSettingsManager super-class so that
all HostBasedTests can inherit it.

For now, I'm using the same set of report types as an image build.  We
may want to adjust that in the future.

Signed-off-by: Jake Garver <jake@nvidia.com>
Tested-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeshua Smith <jeshuas@nvidia.com>