fix: install dummy values if CvmBoardInfo is not available
We still need to install gNVIDIACvmEepromProtocolGuid even if Cvm data is not
available in the hoblist. As a fix, we will treat silicon platforms in this
situation simliarly to simulation platforms - by populating dummy data.
This will satisfy the depex on the SMBIOS driver and allow tables to be installed

Signed-off-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>