fix: RNDIS can not receive data in ready-to-boot callback
Address the issue of receiving data in read-to-boot callback. In RNDIS
driver, a periodic timer callback is used to receive network data in
background. However, this callback will not be called during ready-to-boot
callback because of the same TPL level. Remove this design and call receiver
function directly from SNP receive function and fix this issue.
Address the CPU exception when BMC is reset suddenly.
Introduce USB NIC info protocol to RNDIS driver in order to provide the
interface for HostInterfaceLib to get USB NIC MAC address.

Signed-off-by: Nickle Wang <nicklew@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>