fix: Allow secure boot enablement on AGX Xavier platform
Currently, the PcdUserPhysicalPresence PCD is set to FALSE to prevent
users from being able to modify secure-boot related settings via the
UEFI Menu. This works well on most platforms/SOCs where StandaloneMm
image is present to manage the variables. On AGX Xavier, which doesn't
have StandaloneMm to manage variables, setting this PCD to FALSE
prevents enrolling default keys and enabling secure-boot via the Key
Enrollment App. To workaround this issue, set the PhysicalPresence Pcd
to TRUE for AGX Xavier Devices to allow the default keys to be enrolled,
and set the PCD to FALSE when the next reboot is done.

Signed-off-by: Girish Mahadevan <gmahadevan@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>