chore: Remove misleading message from customer build
This removes the "Building from tarball" message from edkrepo builds.
The message is misleading.

Also, remove support for calling "stuart_setup" during build prepartion.
This only applied when the root of our workspace was a git repo, but
we've since moved to edkrepo for workspace creation and it loads
submodules for us.  stuart_setup is never called.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: William Watson <wwatson@nvidia.com>