feat: configure debug print level on debug builds
Allow for configuration of the DebugPrintErrorLevel on DEBUG builds
via UEFI setup. This allows the developer to quickly enable more or
less debug without rebuilding or flashing the system. This menu is
hidden in RELEASE mode as it is a development feature.

Based on stopwatch and dp command testing, this change adds less
than 500ms of boot time to the system due to a HOB search and access
on the load of each module.

Signed-off-by: Nick Ramirez <nramirez@nvidia.com>