fix: Handle device-tree nodes with one clock
If the device has only one clock, then it is not mandatory for the
device-tree 'clock-names' property to be specified. The 'clock-names'
for the Tegra PWM device were recently removed in device-tree because
there is only one clock and this is causing an assert. Fix this by
checking to see if the device only has one clock, if it does and the
clock name is not valid, then we assume that that is the correct clock.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>