fix: config TegraPlatformCompatSpec variable as writable
The value of TegraPlatformCompatSpec variable is
generated according to the logic of board FAB and
board SKU in board config file (such as
jetson-agx-orin-devkit.conf). The logic maybe
changed in different release, so the
TegraPlatformCompatSpec maybe changed in
different release as well.

Currently, the TegraPlatformCompatSpec variable
is configured as read-only, this CL change it
as writable.

Signed-off-by: Russell Xiao <russellx@nvidia.com>
Reviewed-by: Jimmy Zhang <jimmzhang@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>