fix: config TegraPlatformCompatSpec variable as writable
The value of TegraPlatformCompatSpec variable is generated according to
the logic of board FAB and board SKU in board config file (such as
jetson-agx-orin-devkit.conf). The logic may be changed in different
release, so the TegraPlatformCompatSpec may be changed in different
release as well.

Currently, the TegraPlatformCompatSpec variable is configured as
read-only.  After this change, it is writable.

Signed-off-by: Russell Xiao <russellx@nvidia.com>
Reviewed-by: Jimmy Zhang <jimmzhang@nvidia.com>