fix: Increase allocation size for base DTB
UEFI applies DT overlay file on base DTB. For this, it allocates double
the size of base DTB.  That is now insufficient.  Many of the nodes have
moved from the base DTB and into overlays.  As a result, the base DTB has
become smaller.

This change increases memory allocation to 4X of base DTB size.

Signed-off-by: Petlozu Pravareshwar <petlozup@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>