fix(TegraVirt): Use a Null ResetSystemLib
ArmVirt.dsc.inc uses ArmVirtPsciResetSystemLib, but it is incompatible
with our DTB.  It expects arm,psci-0.2 and we have arm,psci-1.0.  Long
term, we want to resolve that discrepancy and go back to
ArmVirtPsciResetSystemLib, but for now we can use the Null lib and
ignore reset requests.

ResetSystemRuntimeDxe requires a ResetSystemLib.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>