fix(L4TLauncher): delete BootChainFwStatus before switching Rootfs chain
By default, after fresh flashing, system suppose both rootfs A and
rootfs B are good. Rootfs A/B logic switches to non-current rootfs if
current rootfs slot failed to boot for three times. UEFI executes boot
chain switching only when BootChainFwStatus variable does not exist.
Normally the variable is deleted by background service if boot rootfs
successfully.

Suppose current rootfs is A, non-current rootfs is B. If rootfs B is
somehow corrupted, when switching to rootfs B from A, system tries to
boot rootfs B and failed for three times, then it tries to switch back
rootfs A.

However, the BootChainFwStatus variable is not deleted because rootfs B
boot failed, backgroud service can not delete the variable. UEFI does
not switch back to rootfs A and go on boot to Rootfs A/B logic, Rootfs
A/B logic finds out that the rootfs B is unbootable, sets
BootChainFwNext to switch back to rootfs A and reboot. While UEFI will
go on boot to Rootfs A/B logic and not switch to rootfs A because of the
variable. Rootfs A/B logic will ask for switching back to rootfs A and
reboot. Then system enters infinite reboot.

So, delete the BootChainFwStatus variable before switching the rootfs
slot.

Signed-off-by: Russell Xiao <russellx@nvidia.com>
Reviewed-by: Bob Morgan <bobm@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>