feat: set SPI TIMEOUT to 0 globally instead of adding Quick* APIs
The Quick* APIs that had been used for doing accecces with TIMEOUT
set to 0 are removed, and the delays that used TIMEOUT are removed.

This testing doesn't cover MacronixAsp, so TIMEOUT is renamed to
MX_TIMEOUT and the delay isn't removed in that code to preserve the
prior behavior of TIMEOUT there.

This replaces commit f7d66246c76604130c70f0ba58619fc37545ba84.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>