feat: Add Kconfig configuration to builds
Add support for using kconfig to build UEFI configuration.
The Kconfig file defines all options and rules for validity.
Each platform has a defconfig that sets the overrides for this.
Build system generates a .config file in nvidia-config/<PlatformName>
if it does not already exist that is used by the build systems

kconfig-frontends can be used to edit these with a menu.

Known issue, the UEFI parser seems to print warnings on !if macros that
are not defined even if there is an ifdef guarding it.

Signed-off-by: Jeff Brasen <jbrasen@nvidia.com>