fix: In format strings, correct truncation mistakes
Resolve cppcheck findings related to missing or extraneous "l"
modifiers.  These can result in erroneous output.

In edk2, the "l" is treated as a UINT64, otherwise, the argument is cast
as an int.  Other modifiers, such as "h" and "hh" are ignored.

The "ll" modifier is treated the same as an "l".  While the double "l"
is redundant in edk2, it's not incorrect.  Therefore, these were left as
is.

Signed-off-by: Jake Garver <jake@nvidia.com>