feat: Introduce ASSERT alternatives
Introduce NV_ASSERT_RETURN and NV_ASSERT_EFI_ERROR_RETURN.  These
capture common patterns and encourage proper handling of errors when
debug is disabled.

For example, a common pattern is:

    if (Cond) {
       DEBUG ((DEBUG_ERROR, "Msg"));
       ASSERT (Cond);
       return Status;
    }

This can be written as:

    ASSERT_RETURN (!Cond, return Status, Msg);

Further, the Action (second parameter) is required, which encourages
developers to provide a non-DEBUG action.

The NV_ASSERT_EFI_STATUS_RETURN macro is a similar extension to the
existing NV_ASSERT_EFI_STATUS macro.

These macros may be upstreamed to edk2 in the future.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Jeff Brasen <jbrasen@nvidia.com>