test: Implement MemoryAllocationLib stub
This is a fork of MemoryAllocationLibPosix, which is a unittest library.
It will also allow us to simulate out-of-memory situations and check for
memory leaks.  For now, just the AllocatePool () function has been
mocked.

This implementation can be used as a drop-in replacement for
MemoryAllocationLibPosix.  However, if tests call MockAllocatePool (),
they'll also need to call MemoryAllocationStubLibInit () to reset state.

This work will support AndroidBootDxe unittesting.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Bob Morgan <bobm@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>