feat(Android): Check header version
Before loading an Android image, check the header version and verify
it's a version we support.

The original boot.img header format has been formalized as type 0. Types
1-4 have since been added and they are not backwards compatible.  The
only fields that are consistent between each type are the magic and the
version.

AndroidBootGetVerify () will now operate by first reading enough of the
header to get the magic and version.  The magic verifies we are indeed
looking at a boot.img header.  With the version, we'll read the header
again, but this time we'll know how much to read and how to cast it.

Also, formalize the existing support as support for type 0.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>