feat: Support Redfish Boot Order
Support Redfish BootOrder attribute under
compute system resource.
- Add PlatformRedfishBootDxe to handle boot order change request
from Redfish. The HII created in this driver will not be displayed
in BIOS setup menu. It works in background to serve Redfish
boot order request and apply it to system.
- We can not use boot order in UiApp that is because HiApp is an
application instead of UEFI driver. UiApp will be called only
when user select to enter setup menu. All the boot order functions
will not be called if user does not enter setup menu.
- Fix the memory issue in PlatformRedfishBiosDxe because we should
not free the memory which is allocated by caller.
- Add corresponding drivers to Server DSC and FDF files.

Signed-off-by: Nickle Wang <nicklew@nvidia.com>
Reviewed-by: Thejaswani Putta <tputta@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Tested-by: Nick Ramirez <nramirez@nvidia.com>