fix: Exception in TPM driver
Normally, TPM failure would be detected when Tpm2RequestUseTpm is
called and handle properly.

However, there is a corner case that reading TPM register in
Tpm2RequestUseTpm is okay, but TPM fails to register the successive
commands and does not provide response. But the current code still
attempts to print the response buffer, which triggers exception.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>