feat: support interrupts-extended parsing of DTB
The interrupt parsing code for the DTB helpers is rewritten to support
the interrupts-extended field, as well as to look up the correct number
of #interrupt-cells for the controller that handles the interrupts
rather than hardcoding that number to 3.

A pointer to the compatibility string of the controller is passed
back in the results, to allow callers of the code to determine which
interrupts go to which controllers.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>