feat: I2C mux/switch devices via EFI_I2C_BUS_CONFIGURATION_MANAGEMENT_PROTOCOL
Previously I2C devices behind a mux or switch were not visible to UEFI, and camera
EEPROMs behind a mux would not get probed during boot.

This commit adds support for all pca954x type I2C mux/switches, which the TCA9548A
on the reference e3333 interposer board is compatible with.

This allows the device tree to define an I2c mux using a `compatible = "nxp,pca9547"`
prop, with child nodes inside, and for UEFI to be able to read camera EEPROMS behind
these muxes. It does not yet add support for "i2c-mux-gpio" type muxes, but could
potentially be added later.

Signed-off-by: eh-steve <16373174+eh-steve@users.noreply.github.com>