chore: Fork AndroidBootImgLib.h
Forking edk2's EmbeddedPkg/Include/Library/AndroidBootImgLib.h.  We'll
be adding type 0-4 definitions to it.  Over time, we plan to upstream
these definitions, but it's unclear if/how these can be integrated into
existing edk2 code.

Our Android boot support diverges significantly from edk2's and the
support in edk2 is stale.  Our driver is currently getting the boot
header definition from edk2, but that definition pre-dates the
formalization of the boot.img header.  Rather than take on updating the
edk2 support, which we're not using, we'll fork the header and continue
building our driver independently.

This commit brings in AndroidBootImgLib.h without changes so that we can
track our updates more easily.

Signed-off-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>