fix: Exception in Tcg2Dxe
On platform with TPM, after `reconnect -r` is called, Tpm2Dxe driver
would install to different address. And since CloseEvent() was called in
Tpm2DeviceLib, it did not update the pointer to Tpm2Dxe protocol, and
triggered exception.

This patch fixes the issue by closing the event only when TPM isn't
available, which is when the library destructor is called.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Nickle Wang <nicklew@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>