feat: Enable exception handler during PrePi
This change adds exception handling to PrePi, similar to what we have in
DXE.  In PrePi, we'll use ArmExceptionLib as we do in DXE, but replace
DefaultExceptionHandlerLib with a fork that works in the PrePi
environment.  Primarily, this fork will not depend on the EFI Debug
Image Table, but will instead assume we're in PrePi.

We'll use the load address of the image as the exception's base address.
That gives us locations that match PrePi.map.

To get the load address, we'll use PcdFvBaseAddress, which is now
populated in PrePi.c.  This is an ArmPkg GUID that may be useful in
other places.

Signed-off-by: Jake Garver <jake@nvidia.com>