chore: Update Nuvoton driver to access RTC in runtime
Originally, Nuvoton RTC driver does not directly write to RTC during OS
runtime, but stages the new time difference in "RTC_OFFSET" variable and
applies it to RTC on the next reboot.

This approach does not allow the host to sync RTC time with BMC. Another
flaw is if the host is left running for a long period of time, the RTC
drift would become significant and the RTC_OFFSET would not account for
the drift, so on the next boot, the updated time would be quite off.

This patch changes the Nuvoton RTC driver to read and write directly to
RTC, but it still keeps the performance counter approach to support
higher precision.

Signed-off-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>