feat(NvDisplayControllerDxe): use custom EDID blob for display
Expose an instance of EFI_EDID_OVERRIDE_PROTOCOL to force the UEFI
display driver to use a custom EDID binary blob. This is useful
e.g. for use cases where the attached display device is fixed and
known in compile-time, but does not support reading EDID in runtime.

Additionally, include EFI_EDID_OVERRIDE_ENABLE_HOT_PLUG in the
attributes to force the display output to be activated even when the
driver fails to detect any attached display device.

The included EDID binary blob is taken from:

  VESA Enhanced EDID Standard, Release A, Rev.2 (EDID 1.4); Appendix A
  (Sample EDIDs), Example 2.

Signed-off-by: Jan Bobek <jbobek@nvidia.com>