fix(t194): Fix error with no serial ports configured on T194
The behavior of `UpdateSerialPortInfo` for the T234 differs from that of
the T194. When all serial ports are disabled on the T234, DSDT is not
patched with serial port info and the ConfigurationManager DXE resumes
as normal. When this is the case on the T194, the error bubbles up to
the top of ConfigurationManager's entrypoint, causing many further
configuration tables to not be populated (SMBIOS being one of them).

Signed-off-by: Jared Baur <jaredbaur@fastmail.com>
Tested-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>